+++ tasks/package.js
@@ -43,7 +43,8 @@ module.exports = function(grunt) {
         }
       ],
       options: {
-        gruntLogHeader: false
+        gruntLogHeader: false,
+        mode: true
       }
     });