+++ CHANGELOG.md
@@ -1,5 +1,47 @@
 # CHANGELOG
 
+## v0.9.3 [2016/02/25]
+
+- Improved performance for code style quality checks by limiting file scans.
+- Updated default file patterns for code style quality checks.
+- Added Travis CI test coverage for more versions of Node.js.
+- Fixed Drupal 8 tests by adding a D8-specific composer.json file.
+
+## v0.9.2 [2016/01/27]
+
+- Added `--no-validate` command line option to skip running the `grunt:validate`
+tasks, including PHPCS and eslint.
+- Fixed bug that prevented phpcs and eslint from running
+- Fixed bug that caused scope error for this.name.
+- Updated dependencies. Added npm shrinkwrap file to ensure consistency between
+installations.
+
+## v0.9.1 [2015/12/10]
+
+- Add config option to specify command runner in git hook scripts.
+- Fixed Gruntconfig.json interpretation for disabling tasks and specifying port
+for the serve task.
+- Fixed error when using a theme's validate script.
+- Configured copy operations during build to ensure file modes are preserved.
+
+## v0.9.0 [2015/11/18]
+
+- Added the ability to [define scripts for common project operations](https://github.com/phase2/grunt-drupal-tasks/blob/master/CONFIG.md#project-operations).
+- Added `grunt install` task to [easily install the site cleanly or with an
+imported database](https://github.com/phase2/grunt-drupal-tasks/blob/master/CONFIG.md#install-settings).
+- Added optional [integration with Git Hooks](https://github.com/phase2/grunt-drupal-tasks/blob/master/CONFIG.md#adding-git-hooks)
+to support running Grunt tasks when git operations are run. When enabled, the
+validation task is run against staged code before a commit is made.
+- If index.php is missing from the build destination, then the Drush make task
+is always run (and the "newer" feature is disabled). This facilitates certain VM
+configurations.
+- PHPCS and eslint tasks are skipped if there are no source files.
+
+### Upgrade Notes
+
+- Themes with a validate task configured are excluded by default from the Grunt
+Drupal Tasks validate task.
+
 ## v0.8.0 [2015/07/24]
 
 - Fixed bug in theme proxying feature.